requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
langchain==0.3.26
langchain-google-genai==2.1.8
pypdf==5.8.0
langchain-community==0.3.27
jinja2==3.1.2
faiss-cpu==1.13.2
pinecone-client==6.0.0
langchain-pinecone==0.2.13
chromadb==1.5.0
rank_bm25==0.2.2
pandas==2.2.2
matplotlib==3.9.4
seaborn==0.13.2
langchain-groq==0.3.8
langchain-experimental==0.3.4
python-dotenv==1.0.0
tabulate==0.9.0